Fix get-attributes method in knowledgebase so cls is taken into account
Added an attributes() method to resolvers that returns a list of lispified attributes that the resolver provides. Simplified the resolver API by restricting params to functions. Resolvers apply to a single class (thus _should_resolve takes a cls param). The resolve function takes a cls, attr, and symbol as params. Attributes are given based on cls and symbol.