Always return unicode
Omnibase is running Guile 1.8, which doesn't support unicode characters. Because of this, we were converting all unicode strings to ascii strings using the unidecode library. It turns out that unidecode (converting unicode strings into normal byte strings) is only necessary when generating symbols and synonyms. It's safe for WikipediaBase to return answers in unicode, as START can handle them correctly.

Modified lispify to always return unicode lispified objects. Left the output() function in util as a way of asserting that all strings are unicode strings, but this should be removed before going to production. Removed unidecode as a dependency.