Return strings encoded into numeric HTML entities instead of UTF-8 encoded strings
Modified the WikipediaBase server to return strings encoded with numeric HTML entities. As of Oct '15, we haven't figured out how to correctly decode UTF-8 strings in Allegro CLisp. The output of WikipediaBase is mangled by START. To prevent this, we're encoding into HTML entities as a temprary workaround (to be fixed before deploying to production). Added namedentities as a dependency.