Split infobox construction in two phases with RenderedInfobox
Split infobox into Infobox and RenderedInfobox, where Infobox only
contains markup and RenderedInfobox contains markup and HTML. Split the
construction of these objects into InfoboxBuilder and
RenderedInfoboxBuilder.

This was motivated by performance reasons: getting the HTML of an
infobox was in the critical path of _every_ request. But get-classes
and get-attributes don't need the HTML of an infobox, they just use
the markup. Only get needs the HTML.

It might be argued that a two-step process to get infoboxes is complex.
While this might be true, the performance realities of how START
accesses WikipediaBase (making a lot of get-classes calls, some
get-attributes calls, and only a few get's) makes this change very
important.

This change results in a massive speedup. Performance between
WikipediaBase and Ruby-WikipediaBase is now comparable.